Crate psl_lexer

Public Suffix List Lexer

All this library does is provide methods for fetching and parsing Mozilla's Public Suffix List. You may be interested in higher level libraries like psl.

Examples

extern crate psl_lexer;

use psl_lexer::List;

// Fetch the list from the official URL,
let list = List::fetch()?;

// from your own URL
let list = List::from_url("https://example.com/path/to/public_suffix_list.dat")?;

// or from a local file.
let list = List::from_path("/path/to/public_suffix_list.dat")?;

Re-exports

pub use errors::Result;

pub use errors::Error;

Modules

errors
Errors returned by this library

Structs

List
Stores the public suffix list

Suffix

Enums

Type

Constants

LIST_URL
The official URL of the list

Traits

IntoUrl
Converts a type into a Url object